Remove some self developed algorithms and support other repos rely on Adlik as a third-party library
