--- adlik_serving/apis/amc_task.proto
@@ -1,18 +1,0 @@
-syntax = "proto3";
-
-package adlik.serving;
-option cc_enable_arenas = true;
-
-message AmcTaskReq {
-  uint32 cell_id = 1;
-
-  // map: grid -> bler
-  map<int32, double> blers = 2;
-}
-
-message AmcTaskRsp {
-  uint32 cell_id = 1;
-
-  // map: grid -> delta
-  map<int32, double> deltas = 2;
-}
