--- adlik_serving/apis/grid_task.proto
@@ -1,21 +1,0 @@
-syntax = "proto3";
-
-package adlik.serving;
-option cc_enable_arenas = true;
-
-message Cell {
-  uint64 plmn = 1;
-  uint32 nb_id = 2;
-  uint32 cell_id = 3;
-}
-
-message GridTaskReq {
-  Cell cell = 1;     // server
-  string input = 2;  // input csv path
-  string output = 3; // output csv path
-}
-
-message GridTaskRsp {
-  Cell cell = 1;     // server
-  string output = 2; // output file path
-}
