+++ adlik_serving/framework/manager/BUILD
@@ -2,7 +2,7 @@ load("//adlik_serving:serving.bzl", "serving_cc_library")
 
 package(
     default_visibility = [
-        "//adlik_serving:internal",
+        "//visibility:public",
     ],
 )
