+++ adlik_serving/runtime/ml/algorithm/algorithm_factory.cc
@@ -11,8 +11,9 @@ cub::StatusWrapper AlgorithmFactory::create(const std::string& name,
                                             const std::string& model_dir,
                                             std::unique_ptr<Algorithm>* algo) {
   auto it = creators.find(name);
-  return it != creators.end() ? it->second(model_dir, algo) :
-                                cub::StatusWrapper(cub::Internal, "Not found algorithm creator by name " + name);
+  return it != creators.end() ?
+             it->second(model_dir, algo) :
+             cub::StatusWrapper(cub::Internal, "Not found algorithm creator by name \"" + name + "\"");
 }
 
 void AlgorithmFactory::add(const std::string& name, AlgoCreator creator) {
