--- adlik_serving/runtime/ml/algorithm/proto/amc.proto
@@ -1,12 +1,0 @@
-syntax = "proto3";
-
-package adlik.serving;
-option cc_enable_arenas = true;
-
-message AmcConfig {
-  message Parameter {
-    double nBler_target = 1;
-    double lambda = 2;
-  }
-  map<uint32, Parameter> cell_parameters = 1;
-}
